requirements.txt
pyyaml==6.0.1
Wand==0.6.13